requirements.txt
aiocqhttp==1.2.5
aiofiles==0.5.0
asgiref==3.2.7
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Django==3.0.6
grpcio==1.30.0
grpcio-tools==1.30.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.5.5
httpx==0.12.1
Hypercorn==0.9.5
hyperframe==5.2.0
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
priority==1.3.0
protobuf==3.12.2
PyMySQL==0.9.3
pytz==2020.1
Quart==0.11.5
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
sqlparse==0.3.1
toml==0.10.0
typing-extensions==3.7.4.2
urllib3==1.25.9
Werkzeug==1.0.1
wsproto==0.15.0
